% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ximage-package.R
\docType{data}
\name{logo_a}
\alias{logo_a}
\alias{logo_n}
\title{R logo as an RGB image}
\description{
Data obtained from png package.
}
\details{
\code{logo_n} A matrix of colour values in native raster form.

\code{logo_a} An array of RGB colour values.
}
\examples{
ximage(logo_n, asp = .3)
ximage(logo_a, extent = c(8, 18, 60, 80), add = TRUE)
rect(8, 60, 18, 80)
}
