% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcontour.R
\name{xcontour}
\alias{xcontour}
\title{Contour matrix data and draw}
\usage{
xcontour(x, extent = NULL, ..., add = FALSE)
}
\arguments{
\item{x}{something we can contour}

\item{extent}{optional, numeric xmin,xmax,ymin,ymax}

\item{...}{
  Arguments passed on to \code{\link{ximage}}
  \describe{
    \item{\code{zlim}}{optional, absolute range of data to map colours to (maintains
comparable colours across plots); values outside display as 'na.col';
single-band numeric data only}
    \item{\code{xlab}}{x axis label, empty by default}
    \item{\code{ylab}}{y axis label, empty by default}
    \item{\code{col}}{colours to map single-band data to}
    \item{\code{breaks}}{a set of finite numeric breakpoints for the colours, one more
break than colour (if not, colours are interpolated to fit)}
    \item{\code{alpha}}{optional constant opacity in \verb{[0, 1]} (or vector/matrix,
recycled) applied on top of any existing alpha channel; not supported for
nativeRaster input}
    \item{\code{na.col}}{colour for missing values, default "transparent"}
  }}

\item{add}{add to plot, or start afresh}
}
\value{
nothing, called for its side effect of creating or adding to a plot
}
\description{
Like contour() but to work with \code{\link[=ximage]{ximage()}}
}
\details{
Input may be a matrix or a list from \code{gdal_raster_data()} in the vapour package or from \code{read_ds()}
in the gdalraster package.
}
\examples{
#EPSG:27200 see https://github.com/mdsumner/volcano
ex <- c(2667400, 2668010, 6478700, 6479570)
v <- volcano[nrow(volcano):1, ncol(volcano):1]
ximage(v, extent = ex, asp = 1)

xcontour(v, add = TRUE, extent = ex, col = "white")
xrect(ex, add = TRUE, border = "hotpink", lwd = 5)
}
