% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcontour.R
\name{xcontour}
\alias{xcontour}
\title{A new contour}
\usage{
xcontour(x, extent = NULL, ..., add = FALSE)
}
\arguments{
\item{x}{something we can contour}

\item{extent}{optional, numeric xmin,xmax,ymin,ymax}

\item{...}{
  Arguments passed on to \code{\link[=ximage]{ximage}}
  \describe{
    \item{\code{zlim}}{optional, range of data to set colour map (to maintain absolute colours across multiple plots)}
    \item{\code{xlab}}{x axis label, empty by default}
    \item{\code{ylab}}{y axis lable, empty by default}
    \item{\code{col}}{optional colours to map matrix/array data to}
  }}

\item{add}{add to plot, or start afresh}
}
\value{
nothing, called for its side effect of creating or adding to a plot
}
\description{
To work with \code{\link[=ximage]{ximage()}}
}
\examples{
ex <-  c(2667394, 2668004, 6478902, 6479772)
v <- volcano[nrow(volcano):1, ncol(volcano):1]
ximage(v, extent = ex, asp = 1)
#im <- whatarelief::imagery(extent = ex, projection = "+proj=nzmg +datum=WGS84")
#ximage(im, add = TRUE, extent = ex)
xcontour(v, add = TRUE, extent = ex, col = "white")
}
