% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ximage-package.R
\docType{package}
\name{ximage-package}
\alias{ximage-package}
\title{ximage: Draw Images of Raster Data and Related Adornments}
\description{
Draw images easily, set up a plot with an image, specify where that image should be placed. Image plot by default reflects the index of the image data itself, or can be specified in simple extent terms 'xmin,xmax,ymin,ymax'. Numeric matrices, integer arrays, byte arrays, character arrays, and native rasters are (or will be) supported. A combination of image() and rasterImage() from the 'graphics' package with their good features in one place.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/ximage}
  \item Report bugs at \url{https://github.com/hypertidy/ximage/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Michael D. Sumner \email{mdsumner@gmail.com} [copyright holder]
}

Other contributors:
\itemize{
  \item Chris Toney [contributor]
}

}
\keyword{internal}
