% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrect.R
\name{xrect}
\alias{xrect}
\title{xrect}
\usage{
xrect(x, add = FALSE, ..., asp = 1L)
}
\arguments{
\item{x}{four columns worth of rectangles}

\item{add}{instantiate a plot or add to existing (default is \code{add = FALSE})}

\item{...}{arguments passed to \code{\link[=rect]{rect()}}}

\item{asp}{aspect ratio, defaults to 1}
}
\value{
nothing, called for side effect of creating or adding to a plot
}
\description{
Draw rectangles from four colummns xmin,xmax,ymin,ymax
}
\details{
Calls \code{\link[=rect]{rect()}}, but will instantiate a plot if \code{add = FALSE}.
}
\examples{
xrect(runif(100))
xrect(sort(runif(100)))
xrect(runif(100), col = hcl.colors(25, alpha = seq(.2, .8, length.out = 25)))

example(xcontour, ask = FALSE)
xrect(ex, add = TRUE, lwd = 5, lty = 2)
}
